refactor: split gql schema
