+++ src/routes/graphql/schema-gql/resolvers/user.ts
@@ -1,0 +1,86 @@
+import { FastifyInstance } from 'fastify';
+import { MemberTypeEntity } from '../../../../utils/DB/entities/DBMemberTypes';
+import { PostEntity } from '../../../../utils/DB/entities/DBPosts';
+import { ProfileEntity } from '../../../../utils/DB/entities/DBProfiles';
+import { UserEntity } from '../../../../utils/DB/entities/DBUsers';
+
+export const getUserPosts = async (parent: UserEntity, args: unknown, contextValue: FastifyInstance) => {
+  const postsReq = await contextValue.inject({
+    method: 'GET',
+    url: '/posts',
+  });
+  const posts = postsReq.json<PostEntity[]>();
+  return posts.filter((post) => post.userId === parent.id);
+};
+
+export const getUserProfiles = async (parent: UserEntity, args: unknown, contextValue: FastifyInstance) => {
+  const profilesReq = await contextValue.inject({
+    method: 'GET',
+    url: `/profiles`,
+  });
+  const profiles = profilesReq.json<ProfileEntity[]>();
+  return profiles.filter((profile) => profile.userId === parent.id);
+};
+
+export const getUserProfile = async (parent: UserEntity, args: unknown, contextValue: FastifyInstance) => {
+  const profilesReq = await contextValue.inject({
+    method: 'GET',
+    url: `/profiles`,
+  });
+  const profiles = profilesReq.json<ProfileEntity[]>();
+  return profiles.find((profile) => profile.userId === parent.id) || null;
+};
+
+export const getUserMemberTypes = async (parent: UserEntity, args: unknown, contextValue: FastifyInstance) => {
+  const memberTypesReq = await contextValue.inject({
+    method: 'GET',
+    url: `/member-types`,
+  });
+  const memberTypes = memberTypesReq.json<MemberTypeEntity[]>();
+
+  const profilesReq = await contextValue.inject({
+    method: 'GET',
+    url: `/profiles`,
+  });
+  const profiles = profilesReq.json<ProfileEntity[]>();
+  const userProfile = profiles.find((profile) => profile.userId === parent.id);
+
+  return memberTypes.filter((memberType) => memberType.id === (userProfile ? userProfile.memberTypeId : null));
+};
+
+export const getUserMemberType = async (parent: UserEntity, args: unknown, contextValue: FastifyInstance) => {
+  const memberTypesReq = await contextValue.inject({
+    method: 'GET',
+    url: `/member-types`,
+  });
+  const memberTypes = memberTypesReq.json<MemberTypeEntity[]>();
+
+  const profilesReq = await contextValue.inject({
+    method: 'GET',
+    url: `/profiles`,
+  });
+  const profiles = profilesReq.json<ProfileEntity[]>();
+  const userProfile = profiles.find((profile) => profile.userId === parent.id);
+
+  return memberTypes.find((memberType) => memberType.id === (userProfile ? userProfile.memberTypeId : null)) || null;
+};
+
+export const getUserSubscribedTo = async (parent: UserEntity, args: unknown, contextValue: FastifyInstance) => {
+  const usersReq = await contextValue.inject({
+    method: 'GET',
+    url: '/users',
+  });
+  const users = usersReq.json<UserEntity[]>();
+
+  return users.filter((someUser) => someUser.subscribedToUserIds.includes(parent.id));
+};
+
+export const getSubscribedToUser = async (parent: UserEntity, args: unknown, contextValue: FastifyInstance) => {
+  const usersReq = await contextValue.inject({
+    method: 'GET',
+    url: '/users',
+  });
+  const users = usersReq.json<UserEntity[]>();
+
+  return users.filter((someUser) => parent.subscribedToUserIds.includes(someUser.id));
+};
