+++ src/routes/graphql/schema-gql/types/dataTypes.ts
@@ -1,0 +1,83 @@
+import { GraphQLObjectType, GraphQLNonNull, GraphQLID, GraphQLString, GraphQLInt, GraphQLList } from 'graphql';
+import {
+  getSubscribedToUser,
+  getUserMemberType,
+  getUserMemberTypes,
+  getUserPosts,
+  getUserProfile,
+  getUserProfiles,
+  getUserSubscribedTo,
+} from '../resolvers/user';
+
+export const profileType = new GraphQLObjectType({
+  name: 'Profile',
+  fields: {
+    id: { type: new GraphQLNonNull(GraphQLID) },
+    avatar: { type: new GraphQLNonNull(GraphQLString) },
+    sex: { type: new GraphQLNonNull(GraphQLString) },
+    birthday: { type: new GraphQLNonNull(GraphQLInt) },
+    country: { type: new GraphQLNonNull(GraphQLString) },
+    street: { type: new GraphQLNonNull(GraphQLString) },
+    city: { type: new GraphQLNonNull(GraphQLString) },
+    userId: { type: new GraphQLNonNull(GraphQLID) },
+    memberTypeId: { type: new GraphQLNonNull(GraphQLString) },
+  },
+});
+
+export const memberTypeType = new GraphQLObjectType({
+  name: 'MemberType',
+  fields: {
+    id: { type: new GraphQLNonNull(GraphQLString) },
+    discount: { type: new GraphQLNonNull(GraphQLInt) },
+    monthPostsLimit: { type: new GraphQLNonNull(GraphQLInt) },
+  },
+});
+
+export const postType = new GraphQLObjectType({
+  name: 'Post',
+  fields: {
+    id: { type: new GraphQLNonNull(GraphQLID) },
+    title: { type: new GraphQLNonNull(GraphQLString) },
+    content: { type: new GraphQLNonNull(GraphQLString) },
+    userId: { type: new GraphQLNonNull(GraphQLID) },
+  },
+});
+
+export const userType: GraphQLObjectType = new GraphQLObjectType({
+  name: 'User',
+  fields: () => ({
+    id: { type: new GraphQLNonNull(GraphQLID) },
+    firstName: { type: new GraphQLNonNull(GraphQLString) },
+    lastName: { type: new GraphQLNonNull(GraphQLString) },
+    email: { type: new GraphQLNonNull(GraphQLString) },
+    subscribedToUserIds: { type: new GraphQLNonNull(new GraphQLList(new GraphQLNonNull(GraphQLString))) },
+    posts: {
+      type: new GraphQLNonNull(new GraphQLList(new GraphQLNonNull(postType))),
+      resolve: getUserPosts,
+    },
+    profiles: {
+      type: new GraphQLNonNull(new GraphQLList(new GraphQLNonNull(profileType))),
+      resolve: getUserProfiles,
+    },
+    profile: {
+      type: profileType,
+      resolve: getUserProfile,
+    },
+    memberTypes: {
+      type: new GraphQLNonNull(new GraphQLList(new GraphQLNonNull(memberTypeType))),
+      resolve: getUserMemberTypes,
+    },
+    memberType: {
+      type: memberTypeType,
+      resolve: getUserMemberType,
+    },
+    userSubscribedTo: {
+      type: new GraphQLNonNull(new GraphQLList(new GraphQLNonNull(userType))),
+      resolve: getUserSubscribedTo,
+    },
+    subscribedToUser: {
+      type: new GraphQLNonNull(new GraphQLList(new GraphQLNonNull(userType))),
+      resolve: getSubscribedToUser,
+    },
+  }),
+});
